[irep2] Pin each IEEE arity conjunct and all three shadowed names
The arity guards are three separate out-of-range indexes, and only fma's was
tested: a 0-argument nearbyint() and a 1-argument remainder(x) are as reachable
from a bodiless declaration, and each crashes on its own mutant. The shadowing
pairs now define all three names, since that defect is per name too.

Refs #4715
