--- regression/esbmc/irep2_only_fma_bodiless_arity_fail/main.c
@@ -1,9 +1,0 @@
-#include <assert.h>
-
-double fma(double, double);
-
-int main(void)
-{
-  assert(fma(2.0, 3.0) == 5.0);
-  return 0;
-}
